requirements.txt
allure-pytest==2.15.0
pytest==8.4.1
selenium==4.35.0